관리자 극장 유효성 수정버튼추가
