--- src/main/java/kr/Auditorium/vo/AuditoriumVO.java
@@ -1,36 +1,0 @@
-package kr.Auditorium.vo;
-
-
-public class AuditoriumVO {
-    private int auditoriumID;
-    private int theaterID;
-    private String name;
-    private String Type;
-    // getter, setter
-	public int getAuditoriumID() {
-		return auditoriumID;
-	}
-	public void setAuditoriumID(int auditoriumID) {
-		this.auditoriumID = auditoriumID;
-	}
-	public int getTheaterID() {
-		return theaterID;
-	}
-	public void setTheaterID(int theaterID) {
-		this.theaterID = theaterID;
-	}
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
-	public String getType() {
-		return Type;
-	}
-	public void setType(String type) {
-		Type = type;
-	}
-	
-    
-}
